requirements.txt
gspread==5.12.0
oauth2client==4.1.3
PyYAML==6.0.1
python-dotenv==1.0.0
watchdog==3.0.0
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.24.3
opencv-python==4.8.1.78
openai==1.12.0
anthropic==0.34.0 